requirements.txt
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
visitor==0.1.3
Werkzeug==0.15.6
wincertstore==0.2
WTForms==2.2.1
flask-login==0.4.1
argon2-cffi==19.1.0